+++ CHANGELOG.md
@@ -43,6 +43,7 @@ end
 - [@kddeisz] - Handle string literals that start with `%Q`.
 - [@kddeisz] - Handle question method methods in the predicate of an if with a comment in the body.
 - [@kddeisz] - Fix bare `break` with comments immediately after.
+- [@kddeisz] - Fix for heredocs with comments immediately after the declaration.
 
 # [0.22.0] - 2020-12-08